fix(*): fix serveral configuration bugs
