+++ apps/backend/src/main/java/com/neptune/taskboard/TaskBoardApplication.java
@@ -1,0 +1,13 @@
+package com.neptune.taskboard;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.FilterType;
+
+@SpringBootApplication
+public class TaskBoardApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(TaskBoardApplication.class, args);
+    }
+}
