+++ apps/backend/src/main/java/com/neptune/taskboard/configuration/DatabaseConfiguration.java
@@ -1,0 +1,89 @@
+package com.neptune.taskboard.configuration;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.jdbc.DataSourceBuilder;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.env.Environment;
+import org.springframework.dao.annotation.PersistenceExceptionTranslationPostProcessor;
+import org.springframework.orm.jpa.JpaTransactionManager;
+import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
+import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
+
+import javax.sql.DataSource;
+import java.util.Properties;
+
+@Configuration
+public class DatabaseConfiguration {
+
+    @Autowired
+    Environment env;
+
+    /**
+     * DataSource definition for database connection. Settings are read from
+     * the application.properties file (using the env object).
+     */
+    @Bean
+    public DataSource getDataSource(){
+        DataSourceBuilder<?> dataSourceBuilder = DataSourceBuilder.create();
+        dataSourceBuilder.username(env.getProperty("db.username"));
+        dataSourceBuilder.password(env.getProperty("db.password"));
+        dataSourceBuilder.url(env.getProperty("db.url"));
+        return dataSourceBuilder.build();
+    }
+
+    @Bean
+    public LocalContainerEntityManagerFactoryBean entityManagerFactory() {
+        LocalContainerEntityManagerFactoryBean entityManagerFactory =
+                new LocalContainerEntityManagerFactoryBean();
+
+        entityManagerFactory.setDataSource(getDataSource());
+
+        // Classpath scanning of @Component, @Service, etc annotated class
+        entityManagerFactory.setPackagesToScan(
+                env.getProperty("entitymanager.packagesToScan"));
+
+        // Vendor adapter
+        HibernateJpaVendorAdapter vendorAdapter = new HibernateJpaVendorAdapter();
+        entityManagerFactory.setJpaVendorAdapter(vendorAdapter);
+
+        // Hibernate properties
+        Properties additionalProperties = new Properties();
+        additionalProperties.put(
+                "hibernate.dialect",
+                env.getProperty("hibernate.dialect"));
+        additionalProperties.put(
+                "hibernate.show_sql",
+                env.getProperty("hibernate.show_sql"));
+        additionalProperties.put(
+                "hibernate.hbm2ddl.auto",
+                env.getProperty("hibernate.hbm2ddl.auto"));
+        entityManagerFactory.setJpaProperties(additionalProperties);
+
+        return entityManagerFactory;
+    }
+
+
+    /**
+     * Declare the transaction manager.
+     */
+    @Bean
+    public JpaTransactionManager transactionManager() {
+        JpaTransactionManager transactionManager =
+                new JpaTransactionManager();
+        transactionManager.setEntityManagerFactory(entityManagerFactory().getObject());
+        return transactionManager;
+    }
+
+    /**
+     * PersistenceExceptionTranslationPostProcessor is a bean post processor
+     * which adds an advisor to any bean annotated with Repository so that any
+     * platform-specific exceptions are caught and then rethrown as one
+     * Spring's unchecked data access exceptions (i.e. a subclass of
+     * DataAccessException).
+     */
+    @Bean
+    public PersistenceExceptionTranslationPostProcessor exceptionTranslation() {
+        return new PersistenceExceptionTranslationPostProcessor();
+    }
+}
