+++ apps/backend/src/main/java/com/neptune/taskboard/controller/BoardController.java
@@ -1,0 +1,56 @@
+package com.neptune.taskboard.controller;
+
+import com.neptune.taskboard.entity.Board;
+import com.neptune.taskboard.exception.BoardmasterException;
+import com.neptune.taskboard.service.BoardService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/api/dashboard")
+public class BoardController {
+    @Autowired
+    private BoardService service;
+
+    @PostMapping(produces = {MediaType.APPLICATION_JSON_VALUE})
+    public ResponseEntity<Board> createDashboard(@RequestParam String name) {
+        return new ResponseEntity<>(service.create(name), HttpStatus.CREATED);
+    }
+
+    @GetMapping(path = "/{id}", produces = {MediaType.APPLICATION_JSON_VALUE})
+    public ResponseEntity<Board> getDashboardById(@PathVariable Long id) throws BoardmasterException {
+        return new ResponseEntity<>(service.getDashboard(id), HttpStatus.OK);
+    }
+
+    @GetMapping(produces = {MediaType.APPLICATION_JSON_VALUE})
+    public ResponseEntity<Board> getDashboardByName(@RequestParam String name) throws BoardmasterException {
+        return new ResponseEntity<>(service.getDashboard(name), HttpStatus.OK);
+    }
+
+    @GetMapping(path = "/all", produces = {MediaType.APPLICATION_JSON_VALUE})
+    public ResponseEntity<List<Board>> getAllDashboards() {
+        return new ResponseEntity<>(service.getAllDashboards(), HttpStatus.OK);
+    }
+
+    @DeleteMapping(path = "/{id}", produces = {MediaType.APPLICATION_JSON_VALUE})
+    public ResponseEntity<Board> deleteDashboardById(@PathVariable Long id) throws BoardmasterException {
+        return new ResponseEntity<>(service.deleteDashboard(id), HttpStatus.OK);
+    }
+
+    // Eliminar un Dashboard por nombre
+    @DeleteMapping(produces = {MediaType.APPLICATION_JSON_VALUE})
+    public ResponseEntity<Board> deleteDashboardByName(@RequestParam String name) throws BoardmasterException {
+        return new ResponseEntity<>(service.deleteDashboard(name), HttpStatus.OK);
+    }
+
+    // Cambiar el nombre de un Dashboard
+    @PutMapping(produces = {MediaType.APPLICATION_JSON_VALUE})
+    public ResponseEntity<Board> changeDashboardName(@RequestParam String currentName, @RequestParam String newName) throws BoardmasterException {
+        return new ResponseEntity<>(service.changeDashboardName(newName), HttpStatus.OK);
+    }
+}
