+++ apps/backend/src/main/java/com/neptune/taskboard/controller/StateController.java
@@ -10,7 +10,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/state")
+@RequestMapping("/api/state")
 public class StateController {
 
     @Autowired
