+++ apps/backend/src/main/java/com/neptune/taskboard/entity/Task.java
@@ -21,5 +21,5 @@ public class Task {
 
     @ManyToOne
     @JoinColumn(name = "dashboard_id")
-    private Dashboard dashboard;
+    private Board dashboard;
 }
