--- apps/backend/src/main/java/com/neptune/taskboard/unit/BoardMasterApplication.java
@@ -1,11 +1,0 @@
-package com.neptune.taskboard.unit;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BoardMasterApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(BoardMasterApplication.class, args);
-    }
-}
