--- apps/backend/src/main/java/com/neptune/taskboard/unit/entity/State.java
@@ -1,20 +1,0 @@
-package com.neptune.taskboard.unit.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-
-@Entity
-@Setter
-@Getter
-public class State {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    @Column(unique = true, nullable = false)
-    private String name;
-
-    private String description;
-}
