--- apps/backend/src/main/java/com/neptune/taskboard/unit/entity/Task.java
@@ -1,25 +1,0 @@
-package com.neptune.taskboard.unit.entity;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-@Entity
-@Getter
-@Setter
-public class Task {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    private String name;
-    private String description;
-
-    @ManyToOne
-    @JoinColumn(name = "state_id")
-    private State state;
-
-    @ManyToOne
-    @JoinColumn(name = "dashboard_id")
-    private Dashboard dashboard;
-}
