--- apps/backend/src/main/java/com/neptune/taskboard/unit/exception/BoardmasterException.java
@@ -1,16 +1,0 @@
-package com.neptune.taskboard.unit.exception;
-
-import java.io.Serial;
-
-public class BoardmasterException extends Exception{
-    @Serial
-    private static final long serialVersionUID=1L;
-
-    public BoardmasterException(String message) {
-        super(message);
-    }
-
-    public BoardmasterException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
