--- apps/backend/src/main/java/com/neptune/taskboard/unit/exception/BoardmasterExceptionResponse.java
@@ -1,42 +1,0 @@
-package com.neptune.taskboard.unit.exception;
-
-import java.time.LocalDateTime;
-
-public class BoardmasterExceptionResponse {
-
-    private LocalDateTime timestamp;
-    private String message;
-    private String details;
-
-    public BoardmasterExceptionResponse(LocalDateTime timestamp, String message, String details) {
-        super();
-        this.timestamp = timestamp;
-        this.message = message;
-        this.details = details;
-    }
-
-    // Getters and setters
-    public LocalDateTime getTimestamp() {
-        return timestamp;
-    }
-
-    public void setTimestamp(LocalDateTime timestamp) {
-        this.timestamp = timestamp;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-
-    public String getDetails() {
-        return details;
-    }
-
-    public void setDetails(String details) {
-        this.details = details;
-    }
-}
