--- apps/backend/src/main/java/com/neptune/taskboard/unit/repository/IStateRepository.java
@@ -1,10 +1,0 @@
-package com.neptune.taskboard.unit.repository;
-
-import com.neptune.taskboard.unit.entity.State;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface IStateRepository extends JpaRepository<State, Long> {
-    public Optional<State> findByName(String name);
-}
