+++ apps/backend/src/main/resources/application.properties
@@ -12,4 +12,4 @@ hibernate.show_sql= true
 hibernate.hbm2ddl.auto= update
 
 ## Entity manager config
-entitymanager.packagesToScan = com.boardmaster
+entitymanager.packagesToScan = com.neptune.taskboard
