--- apps/backend/src/test/java/com/neptune/taskboard/e2e/TaskboardE2ETest.java
@@ -1,4 +1,0 @@
-package com.neptune.taskboard.e2e;
-
-public class TaskboardE2ETest {
-}
